fixed output_format in mt_bench template

+++ prepare/tasks/llm_as_judge/model_response_assessment.py
@@ -4,7 +4,7 @@ from unitxt.catalog import add_to_catalog
 add_to_catalog(
     FormTask(
         inputs={"question": "str", "model_output": "str"},
-        outputs={"rating_label": "int"},
+        outputs={"rating_label": "float"},
         metrics=["metrics.spearman"],
     ),
     "tasks.rag.model_response_assessment",
